Restrict `client_addr` on Consul agents to localhost and DNSMasq (#87)
Use `override.hcl` which is lexically after `default.json` (generated by
the `run-consul` script)

`client_addr` is used to control which addresses to bind to in order to
server the HTTP API and DNS for the Consul agent. Rarely do you need to
call a Consul Client's HTTP API so makes sense to restrict it to
localhost only.