Reorder User Data
Currently, Fluentd and Prometheus/whatever time series DB can
potentially be run on Nomad. If the Nomad cluster is unhealthy
or the jobs are not running properly, new nodes will never be
able to bootstrap properly. We rearrange the order of launching
so that these nodes can still bootstrap their respective services.

On the downside, this means that we might not even know that the nodes
are not sending their logs.

Perhaps #44 will help?